Fixes seemlying inactive MCM Options Menu
By default, in Papyrus, we set the selected item index of an
options menu to -1 to indicate there is currently no selection. However,
by the time the item is passed through SKSE and shows up in Scaleform,
the value of the number actually reads 4294967295 or 0xffffffff, which
is -1 as a signed int.

The fix implemented here is to detect for this specific value and
properly set it to -1 as intended.

It's unclear why this particular SKSE bug was never identified. Maybe it
would be better to attempt to patch around this particular SKSE bug.